kommunal to'lovlarda qarzdorlikni ham kiritish qilindi

+++ service/rent/rentPayment.service.js
@@ -1,11 +1,11 @@
 const rentPaymentModel = require("../../model/rent/rentPayment.model");
 const rentBalansModel = require("../../model/rent/rentBalans.model");
-const rentCategoryModel = require("../../model/rent/rentCategory.model");
+const rentClientModel = require("../../model/rent/rentClient.model");
 
 class RentPaymentService {
   async getAll(req, res) {
     const allPayments = await rentPaymentModel
-      .find({ categoryId: req.params.id, method: 0 })
+      .find({ clientId: req.params.id, method: 0 })
       .sort({ createdAt: -1 });
     return allPayments;
   }
@@ -19,41 +19,67 @@ class RentPaymentService {
 
     const balans = await rentBalansModel.findOne();
     if (!balans) throw new Error("Balans topilmadi");
-    const isCategory = await rentCategoryModel.findById(req.params.id);
+    const isClient = await rentClientModel.findById(req.params.id);
 
-    if (!isCategory) throw new Error("Kategoriya topilmadi");
+    if (!isclient) throw new Error("Mijoz topilmadi");
 
-    const payData = {
-      amount: req.body.amount,
-      title: req.body.title,
-      month: req.body.month,
-      year: req.body.year,
-      method: 0,
-      user: req.user.id,
-      categoryId: req.params.id,
-    };
-    if (req.body.text) payData.text = req.body.text;
+    if (req.body.debt === true) {
+      const payData = {
+        amount: req.body.amount,
+        title: req.body.title,
+        month: req.body.month,
+        year: req.body.year,
+        method: 1,
+        user: req.user.id,
+        clientId: req.params.id,
+      };
+      if (req.body.text) payData.text = req.body.text;
 
-    const pay = await rentPaymentModel.create(payData);
-    if (!pay) throw new Error("To'lov yaratishda xatolik yuz berdi");
+      const pay = await rentPaymentModel.create(payData);
+      if (!pay) throw new Error("To'lov yaratishda xatolik yuz berdi");
 
-    await rentCategoryModel.findByIdAndUpdate(
-      req.params.id,
-      {
-        $inc: { amount: pay.amount },
-      },
-      { new: true }
-    );
+      await rentClientModel.findByIdAndUpdate(
+          req.params.id,
+          {
+            $inc: { amount: -pay.amount },
+          },
+          { new: true }
+      );
+      return pay;
+    }
+    if(req.body.debt === false){
+      const payData = {
+        amount: req.body.amount,
+        title: req.body.title,
+        month: req.body.month,
+        year: req.body.year,
+        method: 0,
+        user: req.user.id,
+        clientId: req.params.id,
+      };
+      if (req.body.text) payData.text = req.body.text;
 
-    await rentBalansModel.findByIdAndUpdate(
-      balans._id,
-      {
-        $inc: { amount: pay.amount },
-      },
-      { new: true }
-    );
+      const pay = await rentPaymentModel.create(payData);
+      if (!pay) throw new Error("To'lov yaratishda xatolik yuz berdi");
 
-    return pay;
+      await rentClientModel.findByIdAndUpdate(
+          req.params.id,
+          {
+            $inc: { amount: pay.amount },
+          },
+          { new: true }
+      );
+
+      await rentBalansModel.findByIdAndUpdate(
+          balans._id,
+          {
+            $inc: { amount: pay.amount },
+          },
+          { new: true }
+      );
+
+      return pay;
+    }
   }
 
   /* UPDATE */
@@ -70,44 +96,74 @@ class RentPaymentService {
     if (!pay || !balans) {
       throw new Error("Balans yoki to'lov topilmadi");
     }
-    const balansSum = balans.amount;
-    const paySum = pay.amount;
-    const newPay = req.body.amount - paySum;
-    const total = balansSum + newPay;
+    if (pay.method === 0) {
+      const balansSum = balans.amount;
+      const paySum = pay.amount;
+      const newPay = req.body.amount - paySum;
+      const total = balansSum + newPay;
 
-    if (total < 0 || typeof total !== "number") {
-      throw new Error("Balans - minusga chiqib ketadi !!!");
+      if (total < 0 || typeof total !== "number") {
+        throw new Error("Balans - minusga chiqib ketadi !!!");
+      }
+
+      const upPay = await rentPaymentModel.findByIdAndUpdate(
+          req.params.payId,
+          {
+            amount: req.body.amount,
+            title: req.body.title,
+            text: req.body.text,
+            month: req.body.month,
+            year: req.body.year,
+          },
+          { new: true }
+      );
+
+      await rentClientModel.findByIdAndUpdate(
+          upPay.clientId,
+          {
+            $inc: { amount: upPay.amount - pay.amount },
+          },
+          { new: true }
+      );
+
+      await rentBalansModel.findByIdAndUpdate(
+          balans._id,
+          {
+            $inc: { amount: upPay.amount - pay.amount },
+          },
+          { new: true }
+      );
+      return upPay;
     }
 
-    const upPay = await rentPaymentModel.findByIdAndUpdate(
-      req.params.payId,
-      {
-        amount: req.body.amount,
-        title: req.body.title,
-        text: req.body.text,
-        month: req.body.month,
-        year: req.body.year,
-      },
-      { new: true }
-    );
+    if (pay.method === 1) {
 
-    await rentCategoryModel.findByIdAndUpdate(
-      upPay.categoryId,
-      {
-        $inc: { amount: upPay.amount - pay.amount },
-      },
-      { new: true }
-    );
+      if (total < 0 || typeof total !== "number") {
+        throw new Error("Balans - minusga chiqib ketadi !!!");
+      }
 
-    await rentBalansModel.findByIdAndUpdate(
-      balans._id,
-      {
-        $inc: { amount: upPay.amount - pay.amount },
-      },
-      { new: true }
-    );
+      const upPay = await rentPaymentModel.findByIdAndUpdate(
+          req.params.payId,
+          {
+            amount: req.body.amount,
+            title: req.body.title,
+            text: req.body.text,
+            month: req.body.month,
+            year: req.body.year,
+          },
+          { new: true }
+      );
 
-    return upPay;
+      await rentClientModel.findByIdAndUpdate(
+          upPay.clientId,
+          {
+            //              100 000    -    200 000 = -100 000
+            $inc: { amount: pay.amount - upPay.amount  },
+          },
+          { new: true }
+      );
+      return upPay;
+    }
   }
 
   /* DELETE */
@@ -117,24 +173,37 @@ class RentPaymentService {
 
     if (!pay || !balans) throw new Error("Ma'lumot topilmadi");
 
-    await rentCategoryModel.findByIdAndUpdate(
-      pay.categoryId,
-      {
-        $inc: { amount: -pay.amount },
-      },
-      { new: true }
-    );
+    if (pay.method === 0) {
+      await rentCategoryModel.findByIdAndUpdate(
+          pay.clientId,
+          {
+            $inc: { amount: -pay.amount },
+          },
+          { new: true }
+      );
 
-    await rentBalansModel.findByIdAndUpdate(
-      balans._id,
-      {
-        $inc: { amount: -pay.amount },
-      },
-      { new: true }
-    );
+      await rentBalansModel.findByIdAndUpdate(
+          balans._id,
+          {
+            $inc: { amount: -pay.amount },
+          },
+          { new: true }
+      );
 
-    const deleted = await rentPaymentModel.findByIdAndDelete(pay._id);
-    return deleted;
+      const deleted = await rentPaymentModel.findByIdAndDelete(pay._id);
+      return deleted;
+    }
+    if (pay.method === 1) {
+      await rentCategoryModel.findByIdAndUpdate(
+          pay.clientId,
+          {
+            $inc: { amount: pay.amount },
+          },
+          { new: true }
+      );
+      const deleted = await rentPaymentModel.findByIdAndDelete(pay._id);
+      return deleted;
+    }
   }
 }
